元のファイルのデータを書き込むように変更 読み込みは未対応で、データは エンコード方式(2値*2) 日付(4Byteビッグエンディアン) 元のファイル名(最後がNULL文字)の順

+++ 暗号onVBwithWPF/My Project/Settings.Designer.vb
@@ -80,9 +80,9 @@ Partial Friend NotInheritable Class MySettings
     <Global.System.Configuration.UserScopedSettingAttribute(),  _
      Global.System.Diagnostics.DebuggerNonUserCodeAttribute(),  _
      Global.System.Configuration.DefaultSettingValueAttribute("4")>  _
-    Public Property Bits() As Byte
+    Public Property Bits() As Short
         Get
-            Return CType(Me("Bits"),Byte)
+            Return CType(Me("Bits"),Short)
         End Get
         Set
             Me("Bits") = value
